pyproject.toml
[tool.poetry]
name = "vinted"
version = "0.1.0"
description = ""
authors = ["smatand <andrejsmatana0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
pymongo = "^4.6.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
